@@ -1,32 +1,48 @@
 module k8s.io/kube-openapi
 
-go 1.12
+go 1.21
 
 require (
-	github.com/NYTimes/gziphandler v0.0.0-20170623195520-56545f4a5d46
-	github.com/asaskevich/govalidator v0.0.0-20190424111038-f61b66f89f4a
-	github.com/davecgh/go-spew v1.1.1
-	github.com/emicklei/go-restful v0.0.0-20170410110728-ff4f55a20633
-	github.com/go-openapi/jsonreference v0.19.3
-	github.com/go-openapi/swag v0.19.5
-	github.com/golang/protobuf v1.5.2
-	github.com/google/cel-go v0.9.0
-	github.com/google/gofuzz v1.1.0
-	github.com/google/uuid v1.1.2
-	github.com/googleapis/gnostic v0.5.1
-	github.com/mitchellh/mapstructure v1.1.2
-	github.com/munnerz/goautoneg v0.0.0-20120707110453-a547fc61f48d
-	github.com/onsi/ginkgo v0.0.0-20170829012221-11459a886d9c
-	github.com/onsi/gomega v0.0.0-20170829124025-dcabb60a477c
+	github.com/NYTimes/gziphandler v1.1.1
+	github.com/emicklei/go-restful/v3 v3.11.0
+	github.com/go-openapi/jsonreference v0.20.1
+	github.com/go-openapi/swag v0.23.0
+	github.com/google/gnostic-models v0.6.9
+	github.com/google/go-cmp v0.6.0
+	github.com/google/gofuzz v1.2.0
+	github.com/google/uuid v1.6.0
+	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822
+	github.com/onsi/ginkgo/v2 v2.19.0
+	github.com/onsi/gomega v1.33.1
 	github.com/spf13/pflag v1.0.5
-	github.com/stretchr/testify v1.5.1
-	google.golang.org/genproto v0.0.0-20210831024726-fe130286e0e2
-	google.golang.org/protobuf v1.27.1
-	gopkg.in/yaml.v2 v2.2.8
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
-	k8s.io/gengo v0.0.0-20210813121822-485abfe95c7c
-	k8s.io/klog/v2 v2.2.0
-	k8s.io/utils v0.0.0-20210802155522-efc7438f0176
-	sigs.k8s.io/structured-merge-diff/v4 v4.0.2
-	sigs.k8s.io/yaml v1.2.0
+	github.com/stretchr/testify v1.9.0
+	golang.org/x/tools v0.24.0
+	google.golang.org/protobuf v1.35.1
+	gopkg.in/yaml.v3 v3.0.1
+	k8s.io/gengo/v2 v2.0.0-20240826214909-a7b603a56eb7
+	k8s.io/klog/v2 v2.130.1
+	k8s.io/utils v0.0.0-20240711033017-18e509b52bc8
+	sigs.k8s.io/json v0.0.0-20221116044647-bc3834ca7abd
+	sigs.k8s.io/structured-merge-diff/v4 v4.4.2
+	sigs.k8s.io/yaml v1.4.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/go-logr/logr v1.4.2 // indirect
+	github.com/go-openapi/jsonpointer v0.21.0 // indirect
+	github.com/go-task/slim-sprig/v3 v3.0.0 // indirect
+	github.com/google/pprof v0.0.0-20240727154555-813a5fbdbec8 // indirect
+	github.com/josharian/intern v1.0.0 // indirect
+	github.com/json-iterator/go v1.1.12 // indirect
+	github.com/mailru/easyjson v0.7.7 // indirect
+	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
+	github.com/modern-go/reflect2 v1.0.2 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rogpeppe/go-internal v1.12.0 // indirect
+	golang.org/x/mod v0.20.0 // indirect
+	golang.org/x/net v0.28.0 // indirect
+	golang.org/x/sync v0.8.0 // indirect
+	golang.org/x/sys v0.23.0 // indirect
+	golang.org/x/text v0.17.0 // indirect
 )
