@@ -17,11 +17,12 @@ limitations under the License.
 package schemaconv
 
 import (
-	"io/ioutil"
+	"os"
 	"path/filepath"
 	"testing"
 
-	yaml "gopkg.in/yaml.v2"
+	"github.com/google/go-cmp/cmp"
+	yaml "sigs.k8s.io/yaml/goyaml.v2"
 
 	"k8s.io/kube-openapi/pkg/util/proto"
 	prototesting "k8s.io/kube-openapi/pkg/util/proto/testing"
@@ -83,7 +84,7 @@ func testToSchema(t *testing.T, openAPIPath, expectedNewSchemaPath string) {
 		t.Fatal(err)
 	}
 
-	expect, err := ioutil.ReadFile(expectedNewSchemaPath)
+	expect, err := os.ReadFile(expectedNewSchemaPath)
 	if err != nil {
 		t.Fatalf("Unable to read golden data file %q: %v", expectedNewSchemaPath, err)
 	}
@@ -95,5 +96,71 @@ func testToSchema(t *testing.T, openAPIPath, expectedNewSchemaPath string) {
 			filepath.Join("pkg", "schemaconv", expectedNewSchemaPath),
 		)
 		t.Log("You can then use `git diff` to see the changes.")
+		t.Error(cmp.Diff(string(expect), string(got)))
+	}
+}
+
+func TestFieldLevelAnnotation(t *testing.T) {
+	openAPIPath := filepath.Join("testdata", "field-level-annotation.json")
+	fakeSchema := prototesting.Fake{Path: openAPIPath}
+	s, err := fakeSchema.OpenAPISchema()
+	if err != nil {
+		t.Fatalf("failed to get schema for %s: %v", openAPIPath, err)
+	}
+	models, err := proto.NewOpenAPIData(s)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	ns, err := ToSchema(models)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	_ = ns
+
+	// Test to make sure that MapElementRelationship is populated correctly
+	// after being converted from proto type
+	endpointAddress, ok := ns.FindNamedType("io.k8s.api.core.v1.EndpointAddress")
+	if !ok {
+		t.Fatalf("expected to find EndpointAddress")
+	}
+
+	targetRef, ok := endpointAddress.FindField("targetRef")
+	if !ok {
+		t.Fatalf("expected to find EndpointAddress field 'targetRef'")
+	}
+
+	if targetRef.Type.ElementRelationship == nil {
+		t.Fatalf("expected targetRef MapElementRelationship to be atomic")
+	}
+
+	// Test to make sure that schema.Resolve overrides the ElementRelationship
+	// when asked to resolve a reference
+	resolved, ok := ns.Resolve(targetRef.Type)
+	if !ok {
+		t.Fatalf("failed to resolve targetRef type")
+	}
+
+	if resolved.Map == nil {
+		t.Fatalf("expected to resolve to a Map")
+	}
+
+	if resolved.Map.ElementRelationship != *targetRef.Type.ElementRelationship {
+		t.Fatalf("resolved element relationship not converted")
+	}
+
+	// Make sure our test is actually testing something by ensuring the original
+	// relationship is different from what we are changing it to.
+	targetRefWithoutRelationship := targetRef
+	targetRefWithoutRelationship.Type.ElementRelationship = nil
+
+	originalResolved, ok := ns.Resolve(targetRefWithoutRelationship.Type)
+	if !ok {
+		t.Fatalf("failed to resolve targetRef type")
+	}
+
+	if originalResolved.Map.ElementRelationship == *targetRef.Type.ElementRelationship {
+		t.Fatalf("expected original element relationship to differ from field-level override for test")
 	}
 }
