@@ -16,7 +16,6 @@ package validate
 
 import (
 	"encoding/json"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -113,7 +112,7 @@ func TestJSONSchemaSuite(t *testing.T) {
 		}
 	}()
 
-	files, err := ioutil.ReadDir(jsonSchemaFixturesPath)
+	files, err := os.ReadDir(jsonSchemaFixturesPath)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -131,14 +130,14 @@ func TestJSONSchemaSuite(t *testing.T) {
 				return
 			}
 			t.Log("Running " + specName)
-			b, _ := ioutil.ReadFile(filepath.Join(jsonSchemaFixturesPath, fileName))
+			b, _ := os.ReadFile(filepath.Join(jsonSchemaFixturesPath, fileName))
 			doTestSchemaSuite(t, b)
 		})
 	}
 }
 
 func TestSchemaFixtures(t *testing.T) {
-	files, err := ioutil.ReadDir(schemaFixturesPath)
+	files, err := os.ReadDir(schemaFixturesPath)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -152,7 +151,7 @@ func TestSchemaFixtures(t *testing.T) {
 			t.Parallel()
 			specName := strings.TrimSuffix(fileName, filepath.Ext(fileName))
 			t.Log("Running " + specName)
-			b, _ := ioutil.ReadFile(filepath.Join(schemaFixturesPath, fileName))
+			b, _ := os.ReadFile(filepath.Join(schemaFixturesPath, fileName))
 			doTestSchemaSuite(t, b)
 		})
 	}
@@ -160,7 +159,7 @@ func TestSchemaFixtures(t *testing.T) {
 
 func TestOptionalJSONSchemaSuite(t *testing.T) {
 	jsonOptionalSchemaFixturesPath := filepath.Join(jsonSchemaFixturesPath, "optional")
-	files, err := ioutil.ReadDir(jsonOptionalSchemaFixturesPath)
+	files, err := os.ReadDir(jsonOptionalSchemaFixturesPath)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -178,7 +177,7 @@ func TestOptionalJSONSchemaSuite(t *testing.T) {
 				return
 			}
 			t.Log("Running [optional] " + specName)
-			b, _ := ioutil.ReadFile(filepath.Join(jsonOptionalSchemaFixturesPath, fileName))
+			b, _ := os.ReadFile(filepath.Join(jsonOptionalSchemaFixturesPath, fileName))
 			doTestSchemaSuite(t, b)
 		})
 	}
@@ -187,7 +186,7 @@ func TestOptionalJSONSchemaSuite(t *testing.T) {
 // Further testing with all formats recognized by strfmt
 func TestFormat_JSONSchemaExtended(t *testing.T) {
 	jsonFormatSchemaFixturesPath := filepath.Join(formatFixturesPath)
-	files, err := ioutil.ReadDir(jsonFormatSchemaFixturesPath)
+	files, err := os.ReadDir(jsonFormatSchemaFixturesPath)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -205,7 +204,7 @@ func TestFormat_JSONSchemaExtended(t *testing.T) {
 				return
 			}
 			t.Log("Running [extended formats] " + specName)
-			b, _ := ioutil.ReadFile(filepath.Join(jsonFormatSchemaFixturesPath, fileName))
+			b, _ := os.ReadFile(filepath.Join(jsonFormatSchemaFixturesPath, fileName))
 			doTestSchemaSuite(t, b)
 		})
 	}
@@ -219,7 +218,7 @@ func doTestSchemaSuite(t *testing.T, doc []byte) {
 
 	for _, testDescription := range testDescriptions {
 		b, _ := testDescription.Schema.MarshalJSON()
-		tmpFile, err := ioutil.TempFile(os.TempDir(), "validate-test")
+		tmpFile, err := os.CreateTemp(os.TempDir(), "validate-test")
 		assert.NoError(t, err)
 		_, _ = tmpFile.Write(b)
 		tmpFile.Close()
