@@ -4,7 +4,7 @@ import (
 	"reflect"
 	"testing"
 
-	"k8s.io/gengo/types"
+	"k8s.io/gengo/v2/types"
 )
 
 func TestListTypeMissing(t *testing.T) {
@@ -26,7 +26,7 @@ func TestListTypeMissing(t *testing.T) {
 			t: &types.Type{
 				Kind: types.Struct,
 				Members: []types.Member{
-					types.Member{
+					{
 						Name: "Containers",
 						Type: &types.Type{
 							Kind: types.Slice,
@@ -41,7 +41,7 @@ func TestListTypeMissing(t *testing.T) {
 			t: &types.Type{
 				Kind: types.Struct,
 				Members: []types.Member{
-					types.Member{
+					{
 						Name: "Containers",
 						Type: &types.Type{
 							Kind: types.Slice,
@@ -58,14 +58,14 @@ func TestListTypeMissing(t *testing.T) {
 			t: &types.Type{
 				Kind: types.Struct,
 				Members: []types.Member{
-					types.Member{
+					{
 						Name: "Items",
 						Type: &types.Type{
 							Kind: types.Slice,
 						},
 						CommentLines: []string{"+listType=map"},
 					},
-					types.Member{
+					{
 						Name:     "ListMeta",
 						Embedded: true,
 						Type: &types.Type{
@@ -82,13 +82,13 @@ func TestListTypeMissing(t *testing.T) {
 			t: &types.Type{
 				Kind: types.Struct,
 				Members: []types.Member{
-					types.Member{
+					{
 						Name: "Items",
 						Type: &types.Type{
 							Kind: types.Slice,
 						},
 					},
-					types.Member{
+					{
 						Name:     "ListMeta",
 						Embedded: true,
 						Type: &types.Type{
@@ -105,7 +105,7 @@ func TestListTypeMissing(t *testing.T) {
 			t: &types.Type{
 				Kind: types.Struct,
 				Members: []types.Member{
-					types.Member{
+					{
 						Name: "Items",
 						Type: &types.Type{
 							Kind: types.Slice,
@@ -115,12 +115,29 @@ func TestListTypeMissing(t *testing.T) {
 			},
 			expected: []string{"Items"},
 		},
+
+		{
+			name: "a byte-slice field without annotation should pass validation",
+			t: &types.Type{
+				Kind: types.Struct,
+				Members: []types.Member{
+					{
+						Name: "ByteSliceField",
+						Type: &types.Type{
+							Kind: types.Slice,
+							Elem: types.Byte,
+						},
+					},
+				},
+			},
+			expected: []string{},
+		},
 	}
 
 	rule := &ListTypeMissing{}
 	for _, tc := range tcs {
 		if violations, _ := rule.Validate(tc.t); !reflect.DeepEqual(violations, tc.expected) {
-			t.Errorf("unexpected validation result: test name %v, want: %v, got: %v",
+			t.Errorf("unexpected validation result: test name %q, want: %v, got: %v",
 				tc.name, tc.expected, violations)
 		}
 	}
