@@ -156,31 +156,33 @@ func PropertyNotAllowed(name, in, key string) *Validation {
 }
 
 // TooFewProperties an error for an object with too few properties
-func TooFewProperties(name, in string, n int64) *Validation {
-	msg := fmt.Sprintf(tooFewProperties, name, in, n)
+func TooFewProperties(name, in string, minProperties, size int64) *Validation {
+	msg := fmt.Sprintf(tooFewProperties, name, in, minProperties)
 	if in == "" {
-		msg = fmt.Sprintf(tooFewPropertiesNoIn, name, n)
+		msg = fmt.Sprintf(tooFewPropertiesNoIn, name, minProperties)
 	}
 	return &Validation{
 		code:    TooFewPropertiesCode,
 		Name:    name,
 		In:      in,
-		Value:   n,
+		Value:   size,
+		Valid:   minProperties,
 		message: msg,
 	}
 }
 
 // TooManyProperties an error for an object with too many properties
-func TooManyProperties(name, in string, n int64) *Validation {
-	msg := fmt.Sprintf(tooManyProperties, name, in, n)
+func TooManyProperties(name, in string, maxProperties, size int64) *Validation {
+	msg := fmt.Sprintf(tooManyProperties, name, in, maxProperties)
 	if in == "" {
-		msg = fmt.Sprintf(tooManyPropertiesNoIn, name, n)
+		msg = fmt.Sprintf(tooManyPropertiesNoIn, name, maxProperties)
 	}
 	return &Validation{
 		code:    TooManyPropertiesCode,
 		Name:    name,
 		In:      in,
-		Value:   n,
+		Value:   size,
+		Valid:   maxProperties,
 		message: msg,
 	}
 }
@@ -279,6 +281,7 @@ func TooManyItems(name, in string, max int64, value interface{}) *Validation {
 		Name:    name,
 		In:      in,
 		Value:   value,
+		Valid:   max,
 		message: msg,
 	}
 }
@@ -294,6 +297,7 @@ func TooFewItems(name, in string, min int64, value interface{}) *Validation {
 		Name:    name,
 		In:      in,
 		Value:   value,
+		Valid:   min,
 		message: msg,
 	}
 }
@@ -513,6 +517,7 @@ func TooLong(name, in string, max int64, value interface{}) *Validation {
 		Name:    name,
 		In:      in,
 		Value:   value,
+		Valid:   max,
 		message: msg,
 	}
 }
@@ -531,6 +536,7 @@ func TooShort(name, in string, min int64, value interface{}) *Validation {
 		Name:    name,
 		In:      in,
 		Value:   value,
+		Valid:   min,
 		message: msg,
 	}
 }
