@@ -18,6 +18,8 @@ import (
 	"encoding/json"
 
 	"github.com/go-openapi/swag"
+	"k8s.io/kube-openapi/pkg/internal"
+	jsonv2 "k8s.io/kube-openapi/pkg/internal/third_party/go-json-experiment/json"
 )
 
 const (
@@ -41,6 +43,9 @@ type Header struct {
 
 // MarshalJSON marshal this to JSON
 func (h Header) MarshalJSON() ([]byte, error) {
+	if internal.UseOptimizedJSONMarshaling {
+		return internal.DeterministicMarshal(h)
+	}
 	b1, err := json.Marshal(h.CommonValidations)
 	if err != nil {
 		return nil, err
@@ -53,11 +58,33 @@ func (h Header) MarshalJSON() ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-	return swag.ConcatJSON(b1, b2, b3), nil
+	b4, err := json.Marshal(h.VendorExtensible)
+	if err != nil {
+		return nil, err
+	}
+	return swag.ConcatJSON(b1, b2, b3, b4), nil
+}
+
+func (h Header) MarshalNextJSON(opts jsonv2.MarshalOptions, enc *jsonv2.Encoder) error {
+	var x struct {
+		CommonValidations commonValidationsOmitZero `json:",inline"`
+		SimpleSchema      simpleSchemaOmitZero      `json:",inline"`
+		Extensions
+		HeaderProps
+	}
+	x.CommonValidations = commonValidationsOmitZero(h.CommonValidations)
+	x.SimpleSchema = simpleSchemaOmitZero(h.SimpleSchema)
+	x.Extensions = internal.SanitizeExtensions(h.Extensions)
+	x.HeaderProps = h.HeaderProps
+	return opts.MarshalNext(enc, x)
 }
 
 // UnmarshalJSON unmarshals this header from JSON
 func (h *Header) UnmarshalJSON(data []byte) error {
+	if internal.UseOptimizedJSONUnmarshaling {
+		return jsonv2.Unmarshal(data, h)
+	}
+
 	if err := json.Unmarshal(data, &h.CommonValidations); err != nil {
 		return err
 	}
@@ -68,4 +95,24 @@ func (h *Header) UnmarshalJSON(data []byte) error {
 		return err
 	}
 	return json.Unmarshal(data, &h.HeaderProps)
+}
+
+func (h *Header) UnmarshalNextJSON(opts jsonv2.UnmarshalOptions, dec *jsonv2.Decoder) error {
+	var x struct {
+		CommonValidations
+		SimpleSchema
+		Extensions
+		HeaderProps
+	}
+
+	if err := opts.UnmarshalNext(dec, &x); err != nil {
+		return err
+	}
+
+	h.CommonValidations = x.CommonValidations
+	h.SimpleSchema = x.SimpleSchema
+	h.Extensions = internal.SanitizeExtensions(x.Extensions)
+	h.HeaderProps = x.HeaderProps
+
+	return nil
 }
