@@ -18,6 +18,8 @@ import (
 	"encoding/json"
 
 	"github.com/go-openapi/swag"
+	"k8s.io/kube-openapi/pkg/internal"
+	jsonv2 "k8s.io/kube-openapi/pkg/internal/third_party/go-json-experiment/json"
 )
 
 // ParamProps describes the specific attributes of an operation parameter
@@ -34,30 +36,46 @@ type ParamProps struct {
 	AllowEmptyValue bool    `json:"allowEmptyValue,omitempty"`
 }
 
+// Marshaling structure only, always edit along with corresponding
+// struct (or compilation will fail).
+type paramPropsOmitZero struct {
+	Description     string  `json:"description,omitempty"`
+	Name            string  `json:"name,omitempty"`
+	In              string  `json:"in,omitempty"`
+	Required        bool    `json:"required,omitzero"`
+	Schema          *Schema `json:"schema,omitzero"`
+	AllowEmptyValue bool    `json:"allowEmptyValue,omitzero"`
+}
+
 // Parameter a unique parameter is defined by a combination of a [name](#parameterName) and [location](#parameterIn).
 //
 // There are five possible parameter types.
 // * Path - Used together with [Path Templating](#pathTemplating), where the parameter value is actually part
-//   of the operation's URL. This does not include the host or base path of the API. For example, in `/items/{itemId}`,
-//   the path parameter is `itemId`.
+//
+//	of the operation's URL. This does not include the host or base path of the API. For example, in `/items/{itemId}`,
+//	the path parameter is `itemId`.
+//
 // * Query - Parameters that are appended to the URL. For example, in `/items?id=###`, the query parameter is `id`.
 // * Header - Custom headers that are expected as part of the request.
 // * Body - The payload that's appended to the HTTP request. Since there can only be one payload, there can only be
-//   _one_ body parameter. The name of the body parameter has no effect on the parameter itself and is used for
-//   documentation purposes only. Since Form parameters are also in the payload, body and form parameters cannot exist
-//   together for the same operation.
+//
+//	_one_ body parameter. The name of the body parameter has no effect on the parameter itself and is used for
+//	documentation purposes only. Since Form parameters are also in the payload, body and form parameters cannot exist
+//	together for the same operation.
+//
 // * Form - Used to describe the payload of an HTTP request when either `application/x-www-form-urlencoded` or
-//   `multipart/form-data` are used as the content type of the request (in Swagger's definition,
-//   the [`consumes`](#operationConsumes) property of an operation). This is the only parameter type that can be used
-//   to send files, thus supporting the `file` type. Since form parameters are sent in the payload, they cannot be
-//   declared together with a body parameter for the same operation. Form parameters have a different format based on
-//   the content-type used (for further details, consult http://www.w3.org/TR/html401/interact/forms.html#h-17.13.4).
-//   * `application/x-www-form-urlencoded` - Similar to the format of Query parameters but as a payload.
-//   For example, `foo=1&bar=swagger` - both `foo` and `bar` are form parameters. This is normally used for simple
-//   parameters that are being transferred.
-//   * `multipart/form-data` - each parameter takes a section in the payload with an internal header.
-//   For example, for the header `Content-Disposition: form-data; name="submit-name"` the name of the parameter is
-//   `submit-name`. This type of form parameters is more commonly used for file transfers.
+//
+//	`multipart/form-data` are used as the content type of the request (in Swagger's definition,
+//	the [`consumes`](#operationConsumes) property of an operation). This is the only parameter type that can be used
+//	to send files, thus supporting the `file` type. Since form parameters are sent in the payload, they cannot be
+//	declared together with a body parameter for the same operation. Form parameters have a different format based on
+//	the content-type used (for further details, consult http://www.w3.org/TR/html401/interact/forms.html#h-17.13.4).
+//	* `application/x-www-form-urlencoded` - Similar to the format of Query parameters but as a payload.
+//	For example, `foo=1&bar=swagger` - both `foo` and `bar` are form parameters. This is normally used for simple
+//	parameters that are being transferred.
+//	* `multipart/form-data` - each parameter takes a section in the payload with an internal header.
+//	For example, for the header `Content-Disposition: form-data; name="submit-name"` the name of the parameter is
+//	`submit-name`. This type of form parameters is more commonly used for file transfers.
 //
 // For more information: http://goo.gl/8us55a#parameterObject
 type Parameter struct {
@@ -70,6 +88,10 @@ type Parameter struct {
 
 // UnmarshalJSON hydrates this items instance with the data from JSON
 func (p *Parameter) UnmarshalJSON(data []byte) error {
+	if internal.UseOptimizedJSONUnmarshaling {
+		return jsonv2.Unmarshal(data, p)
+	}
+
 	if err := json.Unmarshal(data, &p.CommonValidations); err != nil {
 		return err
 	}
@@ -85,8 +107,31 @@ func (p *Parameter) UnmarshalJSON(data []byte) error {
 	return json.Unmarshal(data, &p.ParamProps)
 }
 
+func (p *Parameter) UnmarshalNextJSON(opts jsonv2.UnmarshalOptions, dec *jsonv2.Decoder) error {
+	var x struct {
+		CommonValidations
+		SimpleSchema
+		Extensions
+		ParamProps
+	}
+	if err := opts.UnmarshalNext(dec, &x); err != nil {
+		return err
+	}
+	if err := p.Refable.Ref.fromMap(x.Extensions); err != nil {
+		return err
+	}
+	p.CommonValidations = x.CommonValidations
+	p.SimpleSchema = x.SimpleSchema
+	p.Extensions = internal.SanitizeExtensions(x.Extensions)
+	p.ParamProps = x.ParamProps
+	return nil
+}
+
 // MarshalJSON converts this items object to JSON
 func (p Parameter) MarshalJSON() ([]byte, error) {
+	if internal.UseOptimizedJSONMarshaling {
+		return internal.DeterministicMarshal(p)
+	}
 	b1, err := json.Marshal(p.CommonValidations)
 	if err != nil {
 		return nil, err
@@ -108,4 +153,20 @@ func (p Parameter) MarshalJSON() ([]byte, error) {
 		return nil, err
 	}
 	return swag.ConcatJSON(b3, b1, b2, b4, b5), nil
+}
+
+func (p Parameter) MarshalNextJSON(opts jsonv2.MarshalOptions, enc *jsonv2.Encoder) error {
+	var x struct {
+		CommonValidations commonValidationsOmitZero `json:",inline"`
+		SimpleSchema      simpleSchemaOmitZero      `json:",inline"`
+		ParamProps        paramPropsOmitZero        `json:",inline"`
+		Ref               string                    `json:"$ref,omitempty"`
+		Extensions
+	}
+	x.CommonValidations = commonValidationsOmitZero(p.CommonValidations)
+	x.SimpleSchema = simpleSchemaOmitZero(p.SimpleSchema)
+	x.Extensions = internal.SanitizeExtensions(p.Extensions)
+	x.ParamProps = paramPropsOmitZero(p.ParamProps)
+	x.Ref = p.Refable.Ref.String()
+	return opts.MarshalNext(enc, x)
 }
