@@ -19,6 +19,8 @@ import (
 	"fmt"
 
 	"github.com/go-openapi/swag"
+	"k8s.io/kube-openapi/pkg/internal"
+	jsonv2 "k8s.io/kube-openapi/pkg/internal/third_party/go-json-experiment/json"
 )
 
 // Swagger this is the root document object for the API specification.
@@ -33,6 +35,9 @@ type Swagger struct {
 
 // MarshalJSON marshals this swagger structure to json
 func (s Swagger) MarshalJSON() ([]byte, error) {
+	if internal.UseOptimizedJSONMarshaling {
+		return internal.DeterministicMarshal(s)
+	}
 	b1, err := json.Marshal(s.SwaggerProps)
 	if err != nil {
 		return nil, err
@@ -44,8 +49,22 @@ func (s Swagger) MarshalJSON() ([]byte, error) {
 	return swag.ConcatJSON(b1, b2), nil
 }
 
+// MarshalJSON marshals this swagger structure to json
+func (s Swagger) MarshalNextJSON(opts jsonv2.MarshalOptions, enc *jsonv2.Encoder) error {
+	var x struct {
+		Extensions
+		SwaggerProps
+	}
+	x.Extensions = internal.SanitizeExtensions(s.Extensions)
+	x.SwaggerProps = s.SwaggerProps
+	return opts.MarshalNext(enc, x)
+}
+
 // UnmarshalJSON unmarshals a swagger spec from json
 func (s *Swagger) UnmarshalJSON(data []byte) error {
+	if internal.UseOptimizedJSONUnmarshaling {
+		return jsonv2.Unmarshal(data, s)
+	}
 	var sw Swagger
 	if err := json.Unmarshal(data, &sw.SwaggerProps); err != nil {
 		return err
@@ -54,6 +73,23 @@ func (s *Swagger) UnmarshalJSON(data []byte) error {
 		return err
 	}
 	*s = sw
+	return nil
+}
+
+func (s *Swagger) UnmarshalNextJSON(opts jsonv2.UnmarshalOptions, dec *jsonv2.Decoder) error {
+	// Note: If you're willing to make breaking changes, it is possible to
+	// optimize this and other usages of this pattern:
+	// https://github.com/kubernetes/kube-openapi/pull/319#discussion_r983165948
+	var x struct {
+		Extensions
+		SwaggerProps
+	}
+
+	if err := opts.UnmarshalNext(dec, &x); err != nil {
+		return err
+	}
+	s.Extensions = internal.SanitizeExtensions(x.Extensions)
+	s.SwaggerProps = x.SwaggerProps
 	return nil
 }
 
@@ -96,6 +132,9 @@ var jsFalse = []byte("false")
 
 // MarshalJSON convert this object to JSON
 func (s SchemaOrBool) MarshalJSON() ([]byte, error) {
+	if internal.UseOptimizedJSONMarshaling {
+		return internal.DeterministicMarshal(s)
+	}
 	if s.Schema != nil {
 		return json.Marshal(s.Schema)
 	}
@@ -106,21 +145,57 @@ func (s SchemaOrBool) MarshalJSON() ([]byte, error) {
 	return jsTrue, nil
 }
 
+// MarshalJSON convert this object to JSON
+func (s SchemaOrBool) MarshalNextJSON(opts jsonv2.MarshalOptions, enc *jsonv2.Encoder) error {
+	if s.Schema != nil {
+		return opts.MarshalNext(enc, s.Schema)
+	}
+
+	if s.Schema == nil && !s.Allows {
+		return enc.WriteToken(jsonv2.False)
+	}
+	return enc.WriteToken(jsonv2.True)
+}
+
 // UnmarshalJSON converts this bool or schema object from a JSON structure
 func (s *SchemaOrBool) UnmarshalJSON(data []byte) error {
+	if internal.UseOptimizedJSONUnmarshaling {
+		return jsonv2.Unmarshal(data, s)
+	}
+
 	var nw SchemaOrBool
-	if len(data) >= 4 {
-		if data[0] == '{' {
-			var sch Schema
-			if err := json.Unmarshal(data, &sch); err != nil {
-				return err
-			}
-			nw.Schema = &sch
+	if len(data) > 0 && data[0] == '{' {
+		var sch Schema
+		if err := json.Unmarshal(data, &sch); err != nil {
+			return err
 		}
-		nw.Allows = !(data[0] == 'f' && data[1] == 'a' && data[2] == 'l' && data[3] == 's' && data[4] == 'e')
+		nw.Schema = &sch
+		nw.Allows = true
+	} else {
+		json.Unmarshal(data, &nw.Allows)
 	}
 	*s = nw
 	return nil
+}
+
+func (s *SchemaOrBool) UnmarshalNextJSON(opts jsonv2.UnmarshalOptions, dec *jsonv2.Decoder) error {
+	switch k := dec.PeekKind(); k {
+	case '{':
+		err := opts.UnmarshalNext(dec, &s.Schema)
+		if err != nil {
+			return err
+		}
+		s.Allows = true
+		return nil
+	case 't', 'f':
+		err := opts.UnmarshalNext(dec, &s.Allows)
+		if err != nil {
+			return err
+		}
+		return nil
+	default:
+		return fmt.Errorf("expected object or bool, not '%v'", k.String())
+	}
 }
 
 // SchemaOrStringArray represents a schema or a string array
@@ -131,6 +206,9 @@ type SchemaOrStringArray struct {
 
 // MarshalJSON converts this schema object or array into JSON structure
 func (s SchemaOrStringArray) MarshalJSON() ([]byte, error) {
+	if internal.UseOptimizedJSONMarshaling {
+		return internal.DeterministicMarshal(s)
+	}
 	if len(s.Property) > 0 {
 		return json.Marshal(s.Property)
 	}
@@ -140,8 +218,23 @@ func (s SchemaOrStringArray) MarshalJSON() ([]byte, error) {
 	return []byte("null"), nil
 }
 
+// MarshalJSON converts this schema object or array into JSON structure
+func (s SchemaOrStringArray) MarshalNextJSON(opts jsonv2.MarshalOptions, enc *jsonv2.Encoder) error {
+	if len(s.Property) > 0 {
+		return opts.MarshalNext(enc, s.Property)
+	}
+	if s.Schema != nil {
+		return opts.MarshalNext(enc, s.Schema)
+	}
+	return enc.WriteToken(jsonv2.Null)
+}
+
 // UnmarshalJSON converts this schema object or array from a JSON structure
 func (s *SchemaOrStringArray) UnmarshalJSON(data []byte) error {
+	if internal.UseOptimizedJSONUnmarshaling {
+		return jsonv2.Unmarshal(data, s)
+	}
+
 	var first byte
 	if len(data) > 1 {
 		first = data[0]
@@ -161,6 +254,18 @@ func (s *SchemaOrStringArray) UnmarshalJSON(data []byte) error {
 	}
 	*s = nw
 	return nil
+}
+
+func (s *SchemaOrStringArray) UnmarshalNextJSON(opts jsonv2.UnmarshalOptions, dec *jsonv2.Decoder) error {
+	switch dec.PeekKind() {
+	case '{':
+		return opts.UnmarshalNext(dec, &s.Schema)
+	case '[':
+		return opts.UnmarshalNext(dec, &s.Property)
+	default:
+		_, err := dec.ReadValue()
+		return err
+	}
 }
 
 // Definitions contains the models explicitly defined in this spec
@@ -193,6 +298,10 @@ func (s StringOrArray) Contains(value string) bool {
 
 // UnmarshalJSON unmarshals this string or array object from a JSON array or JSON string
 func (s *StringOrArray) UnmarshalJSON(data []byte) error {
+	if internal.UseOptimizedJSONUnmarshaling {
+		return jsonv2.Unmarshal(data, s)
+	}
+
 	var first byte
 	if len(data) > 1 {
 		first = data[0]
@@ -220,6 +329,23 @@ func (s *StringOrArray) UnmarshalJSON(data []byte) error {
 		return nil
 	default:
 		return fmt.Errorf("only string or array is allowed, not %T", single)
+	}
+}
+
+func (s *StringOrArray) UnmarshalNextJSON(opts jsonv2.UnmarshalOptions, dec *jsonv2.Decoder) error {
+	switch k := dec.PeekKind(); k {
+	case '[':
+		*s = StringOrArray{}
+		return opts.UnmarshalNext(dec, (*[]string)(s))
+	case '"':
+		*s = StringOrArray{""}
+		return opts.UnmarshalNext(dec, &(*s)[0])
+	case 'n':
+		// Throw out null token
+		_, _ = dec.ReadToken()
+		return nil
+	default:
+		return fmt.Errorf("expected string or array, not '%v'", k.String())
 	}
 }
 
@@ -256,14 +382,29 @@ func (s *SchemaOrArray) ContainsType(name string) bool {
 
 // MarshalJSON converts this schema object or array into JSON structure
 func (s SchemaOrArray) MarshalJSON() ([]byte, error) {
-	if len(s.Schemas) > 0 {
+	if internal.UseOptimizedJSONMarshaling {
+		return internal.DeterministicMarshal(s)
+	}
+	if s.Schemas != nil {
 		return json.Marshal(s.Schemas)
 	}
 	return json.Marshal(s.Schema)
 }
 
+// MarshalJSON converts this schema object or array into JSON structure
+func (s SchemaOrArray) MarshalNextJSON(opts jsonv2.MarshalOptions, enc *jsonv2.Encoder) error {
+	if s.Schemas != nil {
+		return opts.MarshalNext(enc, s.Schemas)
+	}
+	return opts.MarshalNext(enc, s.Schema)
+}
+
 // UnmarshalJSON converts this schema object or array from a JSON structure
 func (s *SchemaOrArray) UnmarshalJSON(data []byte) error {
+	if internal.UseOptimizedJSONUnmarshaling {
+		return jsonv2.Unmarshal(data, s)
+	}
+
 	var nw SchemaOrArray
 	var first byte
 	if len(data) > 1 {
@@ -283,4 +424,16 @@ func (s *SchemaOrArray) UnmarshalJSON(data []byte) error {
 	}
 	*s = nw
 	return nil
+}
+
+func (s *SchemaOrArray) UnmarshalNextJSON(opts jsonv2.UnmarshalOptions, dec *jsonv2.Decoder) error {
+	switch dec.PeekKind() {
+	case '{':
+		return opts.UnmarshalNext(dec, &s.Schema)
+	case '[':
+		return opts.UnmarshalNext(dec, &s.Schemas)
+	default:
+		_, err := dec.ReadValue()
+		return err
+	}
 }
