@@ -17,116 +17,135 @@ limitations under the License.
 package integration
 
 import (
-	"io/ioutil"
 	"os"
 	"os/exec"
+	"path"
 	"path/filepath"
 	"testing"
 
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/onsi/gomega/gexec"
 )
 
 const (
-	testdataDir = "./testdata"
-	testPkgDir  = "k8s.io/kube-openapi/test/integration/testdata"
-	inputDir    = testPkgDir + "/listtype" +
-		"," + testPkgDir + "/maptype" +
-		"," + testPkgDir + "/structtype" +
-		"," + testPkgDir + "/dummytype" +
-		"," + testPkgDir + "/uniontype" +
-		"," + testPkgDir + "/enumtype" +
-		"," + testPkgDir + "/custom" +
-		"," + testPkgDir + "/defaults"
-	outputBase               = "pkg"
-	outputPackage            = "generated"
-	outputBaseFileName       = "openapi_generated"
-	generatedSwaggerFileName = "generated.v2.json"
-	generatedReportFileName  = "generated.v2.report"
-	goldenSwaggerFileName    = "golden.v2.json"
-	goldenReportFileName     = "golden.v2.report"
-	timeoutSeconds           = 5.0
+	headerFilePath             = "../../boilerplate/boilerplate.go.txt"
+	testdataDir                = "./testdata"
+	testPkgRoot                = "k8s.io/kube-openapi/test/integration/testdata"
+	outputPkg                  = testPkgRoot + "/pkg/generated"
+	generatedCodeFileName      = "openapi_generated.go"
+	goldenCodeFilePath         = "pkg/generated/" + generatedCodeFileName
+	generatedSwaggerFileName   = "generated.v2.json"
+	goldenSwaggerFileName      = "golden.v2.json"
+	generatedReportFileName    = "generated.v2.report"
+	goldenReportFileName       = "golden.v2.report"
+	generatedOpenAPIv3FileName = "generated.v3.json"
+	goldenOpenAPIv3Filename    = "golden.v3.json"
+
+	timeoutSeconds = 20.0
 )
+
+var (
+	workingDirectory string
+	tempDir          string
+	terr             error
+	openAPIGenPath   string
+
+	inputDirs = []string{
+		// `path` vs. `filepath` because packages use '/'
+		path.Join(testPkgRoot, "listtype"),
+		path.Join(testPkgRoot, "maptype"),
+		path.Join(testPkgRoot, "structtype"),
+		path.Join(testPkgRoot, "dummytype"),
+		path.Join(testPkgRoot, "uniontype"),
+		path.Join(testPkgRoot, "enumtype"),
+		path.Join(testPkgRoot, "custom"),
+		path.Join(testPkgRoot, "valuevalidation"),
+		path.Join(testPkgRoot, "defaults"),
+	}
+)
+
+func generatedFile(filename string) string { return filepath.Join(tempDir, filename) }
+func testdataFile(filename string) string  { return filepath.Join(testdataDir, filename) }
 
 func TestGenerators(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Integration Test Suite")
 }
 
+var _ = BeforeSuite(func() {
+	// Explicitly manage working directory
+	abs, err := filepath.Abs("")
+	Expect(err).ShouldNot(HaveOccurred())
+	workingDirectory = abs
+
+	// Create a temporary directory for generated swagger files.
+	tempDir, terr = os.MkdirTemp("./", "openapi")
+	Expect(terr).ShouldNot(HaveOccurred())
+
+	// Build the OpenAPI code generator.
+	By("building openapi-gen")
+	binaryPath, berr := gexec.Build("../../cmd/openapi-gen/openapi-gen.go")
+	Expect(berr).ShouldNot(HaveOccurred())
+	openAPIGenPath = binaryPath
+
+	// Run the OpenAPI code generator.
+	By("running openapi-gen")
+	args := append([]string{
+		"--output-dir", tempDir,
+		"--output-pkg", outputPkg,
+		"--output-file", generatedCodeFileName,
+		"--report-filename", generatedFile(generatedReportFileName),
+		"--go-header-file", headerFilePath,
+	}, inputDirs...)
+	command := exec.Command(openAPIGenPath, args...)
+	command.Dir = workingDirectory
+	session, err := gexec.Start(command, GinkgoWriter, GinkgoWriter)
+	Expect(err).ShouldNot(HaveOccurred())
+	Eventually(session, timeoutSeconds).Should(gexec.Exit(0))
+
+	By("writing swagger v2.0")
+	// Create the OpenAPI swagger builder.
+	binaryPath, berr = gexec.Build("./builder/main.go")
+	Expect(berr).ShouldNot(HaveOccurred())
+
+	// Execute the builder, generating an OpenAPI swagger file with definitions.
+	gs := generatedFile(generatedSwaggerFileName)
+	By("writing swagger to " + gs)
+	command = exec.Command(binaryPath, gs)
+	command.Dir = workingDirectory
+	session, err = gexec.Start(command, GinkgoWriter, GinkgoWriter)
+	Expect(err).ShouldNot(HaveOccurred())
+	Eventually(session, timeoutSeconds).Should(gexec.Exit(0))
+
+	By("writing OpenAPI v3.0")
+	// Create the OpenAPI swagger builder.
+	binaryPath, berr = gexec.Build("./builder3/main.go")
+	Expect(berr).ShouldNot(HaveOccurred())
+
+	// Execute the builder, generating an OpenAPI swagger file with definitions.
+	gov3 := generatedFile(generatedOpenAPIv3FileName)
+	By("writing swagger to " + gov3)
+	command = exec.Command(binaryPath, gov3)
+	command.Dir = workingDirectory
+	session, err = gexec.Start(command, GinkgoWriter, GinkgoWriter)
+	Expect(err).ShouldNot(HaveOccurred())
+	Eventually(session, timeoutSeconds).Should(gexec.Exit(0))
+})
+
+var _ = AfterSuite(func() {
+	os.RemoveAll(tempDir)
+	gexec.CleanupBuildArtifacts()
+})
+
 var _ = Describe("Open API Definitions Generation", func() {
-
-	var (
-		workingDirectory string
-		tempDir          string
-		terr             error
-		openAPIGenPath   string
-	)
-
-	testdataFile := func(filename string) string { return filepath.Join(testdataDir, filename) }
-	generatedFile := func(filename string) string { return filepath.Join(tempDir, filename) }
-
-	BeforeSuite(func() {
-		// Explicitly manage working directory
-		abs, err := filepath.Abs("")
-		Expect(err).ShouldNot(HaveOccurred())
-		workingDirectory = abs
-
-		// Create a temporary directory for generated swagger files.
-		tempDir, terr = ioutil.TempDir("./", "openapi")
-		Expect(terr).ShouldNot(HaveOccurred())
-
-		// Build the OpenAPI code generator.
-		By("building openapi-gen")
-		binary_path, berr := gexec.Build("../../cmd/openapi-gen/openapi-gen.go")
-		Expect(berr).ShouldNot(HaveOccurred())
-		openAPIGenPath = binary_path
-
-		// Run the OpenAPI code generator, creating OpenAPIDefinition code
-		// to be compiled into builder.
-		By("processing go idl with openapi-gen")
-		gr := generatedFile(generatedReportFileName)
-		command := exec.Command(openAPIGenPath,
-			"-i", inputDir,
-			"-o", outputBase,
-			"-p", outputPackage,
-			"-O", outputBaseFileName,
-			"-r", gr,
-		)
-		command.Dir = workingDirectory
-		session, err := gexec.Start(command, GinkgoWriter, GinkgoWriter)
-		Expect(err).ShouldNot(HaveOccurred())
-		Eventually(session, timeoutSeconds).Should(gexec.Exit(0))
-
-		By("writing swagger")
-		// Create the OpenAPI swagger builder.
-		binary_path, berr = gexec.Build("./builder/main.go")
-		Expect(berr).ShouldNot(HaveOccurred())
-
-		// Execute the builder, generating an OpenAPI swagger file with definitions.
-		gs := generatedFile(generatedSwaggerFileName)
-		By("writing swagger to " + gs)
-		command = exec.Command(binary_path, gs)
-		command.Dir = workingDirectory
-		session, err = gexec.Start(command, GinkgoWriter, GinkgoWriter)
-		Expect(err).ShouldNot(HaveOccurred())
-		Eventually(session, timeoutSeconds).Should(gexec.Exit(0))
-	})
-
-	AfterSuite(func() {
-		os.RemoveAll(tempDir)
-		gexec.CleanupBuildArtifacts()
-	})
-
-	Describe("openapi-gen --verify", func() {
-		It("Verifies that the existing files are correct", func() {
-			command := exec.Command(openAPIGenPath,
-				"-i", inputDir,
-				"-o", outputBase,
-				"-p", outputPackage,
-				"-O", outputBaseFileName,
-				"-r", testdataFile(goldenReportFileName),
-				"--verify-only",
+	Describe("Validating generated code", func() {
+		It("Generated code should match golden files", func() {
+			// Diff the generated code against the golden code. Exit code should be zero.
+			command := exec.Command(
+				"diff", "-u",
+				goldenCodeFilePath,
+				generatedFile(generatedCodeFileName),
 			)
 			command.Dir = workingDirectory
 			session, err := gexec.Start(command, GinkgoWriter, GinkgoWriter)
@@ -135,13 +154,28 @@ var _ = Describe("Open API Definitions Generation", func() {
 		})
 	})
 
-	Describe("Validating OpenAPI Definition Generation", func() {
+	Describe("Validating OpenAPI V2 Definition Generation", func() {
 		It("Generated OpenAPI swagger definitions should match golden files", func() {
 			// Diff the generated swagger against the golden swagger. Exit code should be zero.
 			command := exec.Command(
-				"diff",
+				"diff", "-u",
 				testdataFile(goldenSwaggerFileName),
 				generatedFile(generatedSwaggerFileName),
+			)
+			command.Dir = workingDirectory
+			session, err := gexec.Start(command, GinkgoWriter, GinkgoWriter)
+			Expect(err).ShouldNot(HaveOccurred())
+			Eventually(session, timeoutSeconds).Should(gexec.Exit(0))
+		})
+	})
+
+	Describe("Validating OpenAPI V3 Definition Generation", func() {
+		It("Generated OpenAPI swagger definitions should match golden files", func() {
+			// Diff the generated swagger against the golden swagger. Exit code should be zero.
+			command := exec.Command(
+				"diff", "-u",
+				testdataFile(goldenOpenAPIv3Filename),
+				generatedFile(generatedOpenAPIv3FileName),
 			)
 			command.Dir = workingDirectory
 			session, err := gexec.Start(command, GinkgoWriter, GinkgoWriter)
@@ -154,7 +188,7 @@ var _ = Describe("Open API Definitions Generation", func() {
 		It("Generated API rule violations should match golden report files", func() {
 			// Diff the generated report against the golden report. Exit code should be zero.
 			command := exec.Command(
-				"diff",
+				"diff", "-u",
 				testdataFile(goldenReportFileName),
 				generatedFile(generatedReportFileName),
 			)
