@@ -1,3 +1,4 @@
+//go:build !ignore_autogenerated
 // +build !ignore_autogenerated
 
 /*
@@ -16,9 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Code generated by openapi-gen.go. DO NOT EDIT.
-
-// This file was autogenerated by openapi-gen. Do not edit it manually!
+// Code generated by openapi-gen. DO NOT EDIT.
 
 package generated
 
@@ -26,36 +25,48 @@ import (
 	common "k8s.io/kube-openapi/pkg/common"
 	spec "k8s.io/kube-openapi/pkg/validation/spec"
 	custom "k8s.io/kube-openapi/test/integration/testdata/custom"
+	defaults "k8s.io/kube-openapi/test/integration/testdata/defaults"
+	enumtype "k8s.io/kube-openapi/test/integration/testdata/enumtype"
+	valuevalidation "k8s.io/kube-openapi/test/integration/testdata/valuevalidation"
+	ptr "k8s.io/utils/ptr"
 )
 
 func GetOpenAPIDefinitions(ref common.ReferenceCallback) map[string]common.OpenAPIDefinition {
 	return map[string]common.OpenAPIDefinition{
-		"k8s.io/kube-openapi/test/integration/testdata/custom.Bac":                      common.EmbedOpenAPIDefinitionIntoV2Extension(custom.Bac{}.OpenAPIV3Definition(), custom.Bac{}.OpenAPIDefinition()),
-		"k8s.io/kube-openapi/test/integration/testdata/custom.Bah":                      schema_test_integration_testdata_custom_Bah(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/custom.Bak":                      custom.Bak{}.OpenAPIDefinition(),
-		"k8s.io/kube-openapi/test/integration/testdata/custom.Bal":                      custom.Bal{}.OpenAPIV3Definition(),
-		"k8s.io/kube-openapi/test/integration/testdata/defaults.Defaulted":              schema_test_integration_testdata_defaults_Defaulted(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/defaults.SubStruct":              schema_test_integration_testdata_defaults_SubStruct(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/dummytype.Bar":                   schema_test_integration_testdata_dummytype_Bar(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/dummytype.Baz":                   schema_test_integration_testdata_dummytype_Baz(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/dummytype.Foo":                   schema_test_integration_testdata_dummytype_Foo(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/dummytype.Waldo":                 schema_test_integration_testdata_dummytype_Waldo(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/enumtype.FruitsBasket":           schema_test_integration_testdata_enumtype_FruitsBasket(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/listtype.AtomicList":             schema_test_integration_testdata_listtype_AtomicList(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/listtype.Item":                   schema_test_integration_testdata_listtype_Item(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/listtype.MapList":                schema_test_integration_testdata_listtype_MapList(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/listtype.SetList":                schema_test_integration_testdata_listtype_SetList(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/listtype.UntypedList":            schema_test_integration_testdata_listtype_UntypedList(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/maptype.AtomicMap":               schema_test_integration_testdata_maptype_AtomicMap(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/maptype.GranularMap":             schema_test_integration_testdata_maptype_GranularMap(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/structtype.AtomicStruct":         schema_test_integration_testdata_structtype_AtomicStruct(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/structtype.ContainedStruct":      schema_test_integration_testdata_structtype_ContainedStruct(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/structtype.DeclaredAtomicStruct": schema_test_integration_testdata_structtype_DeclaredAtomicStruct(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/structtype.GranularStruct":       schema_test_integration_testdata_structtype_GranularStruct(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/uniontype.InlinedUnion":          schema_test_integration_testdata_uniontype_InlinedUnion(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/uniontype.TopLevelUnion":         schema_test_integration_testdata_uniontype_TopLevelUnion(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/uniontype.Union":                 schema_test_integration_testdata_uniontype_Union(ref),
-		"k8s.io/kube-openapi/test/integration/testdata/uniontype.Union2":                schema_test_integration_testdata_uniontype_Union2(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/custom.Bac":                          common.EmbedOpenAPIDefinitionIntoV2Extension(custom.Bac{}.OpenAPIV3Definition(), custom.Bac{}.OpenAPIDefinition()),
+		"k8s.io/kube-openapi/test/integration/testdata/custom.Bah":                          schema_test_integration_testdata_custom_Bah(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/custom.Bak":                          custom.Bak{}.OpenAPIDefinition(),
+		"k8s.io/kube-openapi/test/integration/testdata/custom.Bal":                          custom.Bal{}.OpenAPIV3Definition(),
+		"k8s.io/kube-openapi/test/integration/testdata/custom.FooV3OneOf":                   schema_test_integration_testdata_custom_FooV3OneOf(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/defaults.Defaulted":                  schema_test_integration_testdata_defaults_Defaulted(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/defaults.SubStruct":                  schema_test_integration_testdata_defaults_SubStruct(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/dummytype.Bar":                       schema_test_integration_testdata_dummytype_Bar(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/dummytype.Baz":                       schema_test_integration_testdata_dummytype_Baz(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/dummytype.Foo":                       schema_test_integration_testdata_dummytype_Foo(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/dummytype.StatusError":               schema_test_integration_testdata_dummytype_StatusError(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/dummytype.Waldo":                     schema_test_integration_testdata_dummytype_Waldo(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/enumtype.FruitsBasket":               schema_test_integration_testdata_enumtype_FruitsBasket(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/listtype.AtomicList":                 schema_test_integration_testdata_listtype_AtomicList(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/listtype.Item":                       schema_test_integration_testdata_listtype_Item(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/listtype.MapList":                    schema_test_integration_testdata_listtype_MapList(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/listtype.SetList":                    schema_test_integration_testdata_listtype_SetList(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/listtype.UntypedList":                schema_test_integration_testdata_listtype_UntypedList(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/maptype.AtomicMap":                   schema_test_integration_testdata_maptype_AtomicMap(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/maptype.GranularMap":                 schema_test_integration_testdata_maptype_GranularMap(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/structtype.AtomicStruct":             schema_test_integration_testdata_structtype_AtomicStruct(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/structtype.ContainedStruct":          schema_test_integration_testdata_structtype_ContainedStruct(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/structtype.DeclaredAtomicStruct":     schema_test_integration_testdata_structtype_DeclaredAtomicStruct(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/structtype.FieldLevelOverrideStruct": schema_test_integration_testdata_structtype_FieldLevelOverrideStruct(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/structtype.GranularStruct":           schema_test_integration_testdata_structtype_GranularStruct(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/uniontype.InlinedUnion":              schema_test_integration_testdata_uniontype_InlinedUnion(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/uniontype.TopLevelUnion":             schema_test_integration_testdata_uniontype_TopLevelUnion(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/uniontype.Union":                     schema_test_integration_testdata_uniontype_Union(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/uniontype.Union2":                    schema_test_integration_testdata_uniontype_Union2(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/valuevalidation.Foo":                 schema_test_integration_testdata_valuevalidation_Foo(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/valuevalidation.Foo2":                schema_test_integration_testdata_valuevalidation_Foo2(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/valuevalidation.Foo3":                schema_test_integration_testdata_valuevalidation_Foo3(ref),
+		"k8s.io/kube-openapi/test/integration/testdata/valuevalidation.Foo4":                valuevalidation.Foo4{}.OpenAPIDefinition(),
+		"k8s.io/kube-openapi/test/integration/testdata/valuevalidation.Foo5":                schema_test_integration_testdata_valuevalidation_Foo5(ref),
 	}
 }
 
@@ -65,6 +76,26 @@ func schema_test_integration_testdata_custom_Bah(ref common.ReferenceCallback) c
 			SchemaProps: spec.SchemaProps{
 				Type:   custom.Bah{}.OpenAPISchemaType(),
 				Format: custom.Bah{}.OpenAPISchemaFormat(),
+			},
+		},
+	})
+}
+
+func schema_test_integration_testdata_custom_FooV3OneOf(ref common.ReferenceCallback) common.OpenAPIDefinition {
+	return common.EmbedOpenAPIDefinitionIntoV2Extension(common.OpenAPIDefinition{
+		Schema: spec.Schema{
+			SchemaProps: spec.SchemaProps{
+				Description: "FooV3OneOf has an OpenAPIV3OneOfTypes method",
+				OneOf:       common.GenerateOpenAPIV3OneOfSchema(custom.FooV3OneOf{}.OpenAPIV3OneOfTypes()),
+				Format:      custom.FooV3OneOf{}.OpenAPISchemaFormat(),
+			},
+		},
+	}, common.OpenAPIDefinition{
+		Schema: spec.Schema{
+			SchemaProps: spec.SchemaProps{
+				Description: "FooV3OneOf has an OpenAPIV3OneOfTypes method",
+				Type:        custom.FooV3OneOf{}.OpenAPISchemaType(),
+				Format:      custom.FooV3OneOf{}.OpenAPISchemaFormat(),
 			},
 		},
 	})
@@ -131,6 +162,41 @@ func schema_test_integration_testdata_defaults_Defaulted(ref common.ReferenceCal
 									},
 								},
 							},
+						},
+					},
+					"localSymbolReference": {
+						SchemaProps: spec.SchemaProps{
+							Default: defaults.ConstantValue,
+							Type:    []string{"string"},
+							Format:  "",
+						},
+					},
+					"fullyQualifiedSymbolReference": {
+						SchemaProps: spec.SchemaProps{
+							Default: defaults.ConstantValue,
+							Type:    []string{"string"},
+							Format:  "",
+						},
+					},
+					"externalSymbolReference": {
+						SchemaProps: spec.SchemaProps{
+							Default: enumtype.FruitApple,
+							Type:    []string{"string"},
+							Format:  "",
+						},
+					},
+					"pointerConversionSymbolReference": {
+						SchemaProps: spec.SchemaProps{
+							Default: enumtype.FruitApple,
+							Type:    []string{"string"},
+							Format:  "",
+						},
+					},
+					"defaultedAliasSymbolReference": {
+						SchemaProps: spec.SchemaProps{
+							Default: defaults.ConstantValue,
+							Type:    []string{"string"},
+							Format:  "",
 						},
 					},
 				},
@@ -250,6 +316,33 @@ func schema_test_integration_testdata_dummytype_Foo(ref common.ReferenceCallback
 	}
 }
 
+func schema_test_integration_testdata_dummytype_StatusError(ref common.ReferenceCallback) common.OpenAPIDefinition {
+	return common.OpenAPIDefinition{
+		Schema: spec.Schema{
+			SchemaProps: spec.SchemaProps{
+				Type: []string{"object"},
+				Properties: map[string]spec.Schema{
+					"Code": {
+						SchemaProps: spec.SchemaProps{
+							Default: 0,
+							Type:    []string{"integer"},
+							Format:  "int32",
+						},
+					},
+					"Message": {
+						SchemaProps: spec.SchemaProps{
+							Default: "",
+							Type:    []string{"string"},
+							Format:  "",
+						},
+					},
+				},
+				Required: []string{"Code", "Message"},
+			},
+		},
+	}
+}
+
 func schema_test_integration_testdata_dummytype_Waldo(ref common.ReferenceCallback) common.OpenAPIDefinition {
 	return common.OpenAPIDefinition{
 		Schema: spec.Schema{
@@ -281,15 +374,17 @@ func schema_test_integration_testdata_enumtype_FruitsBasket(ref common.Reference
 	return common.OpenAPIDefinition{
 		Schema: spec.Schema{
 			SchemaProps: spec.SchemaProps{
-				Type: []string{"object"},
+				Description: "FruitsBasket is the type that contains the enum type.",
+				Type:        []string{"object"},
 				Properties: map[string]spec.Schema{
 					"content": {
 						SchemaProps: spec.SchemaProps{
-							Description: "\n\n\nPossible enum values:\n - `apple`: is the Apple\n - `banana`: is the Banana\n - `onigiri`: is the Rice ball that does not seem to belong to a fruits basket but has a long comment that is so long that it spans multiple lines",
+							Description: "Possible enum values:\n - `\"apple\"` is the Apple\n - `\"banana\"` is the Banana\n - `\"onigiri\"` is the Rice ball that does not seem to belong to a fruits basket but has a long comment that is so long that it spans multiple lines",
 							Default:     "",
 							Type:        []string{"string"},
 							Format:      "",
-							Enum:        []interface{}{"apple", "banana", "onigiri"}},
+							Enum:        []interface{}{"apple", "banana", "onigiri"},
+						},
 					},
 					"count": {
 						SchemaProps: spec.SchemaProps{
@@ -613,6 +708,39 @@ func schema_test_integration_testdata_structtype_DeclaredAtomicStruct(ref common
 	}
 }
 
+func schema_test_integration_testdata_structtype_FieldLevelOverrideStruct(ref common.ReferenceCallback) common.OpenAPIDefinition {
+	return common.OpenAPIDefinition{
+		Schema: spec.Schema{
+			SchemaProps: spec.SchemaProps{
+				Type: []string{"object"},
+				Properties: map[string]spec.Schema{
+					"Field": {
+						VendorExtensible: spec.VendorExtensible{
+							Extensions: spec.Extensions{
+								"x-kubernetes-map-type": "atomic",
+							},
+						},
+						SchemaProps: spec.SchemaProps{
+							Default: map[string]interface{}{},
+							Ref:     ref("k8s.io/kube-openapi/test/integration/testdata/structtype.DeclaredAtomicStruct"),
+						},
+					},
+					"OtherField": {
+						SchemaProps: spec.SchemaProps{
+							Default: 0,
+							Type:    []string{"integer"},
+							Format:  "int32",
+						},
+					},
+				},
+				Required: []string{"Field", "OtherField"},
+			},
+		},
+		Dependencies: []string{
+			"k8s.io/kube-openapi/test/integration/testdata/structtype.DeclaredAtomicStruct"},
+	}
+}
+
 func schema_test_integration_testdata_structtype_GranularStruct(ref common.ReferenceCallback) common.OpenAPIDefinition {
 	return common.OpenAPIDefinition{
 		Schema: spec.Schema{
@@ -881,4 +1009,160 @@ func schema_test_integration_testdata_uniontype_Union2(ref common.ReferenceCallb
 			},
 		},
 	}
+}
+
+func schema_test_integration_testdata_valuevalidation_Foo(ref common.ReferenceCallback) common.OpenAPIDefinition {
+	return common.OpenAPIDefinition{
+		Schema: spec.Schema{
+			SchemaProps: spec.SchemaProps{
+				Type:          []string{"object"},
+				MinProperties: ptr.To[int64](1),
+				MaxProperties: ptr.To[int64](5),
+				Properties: map[string]spec.Schema{
+					"StringValue": {
+						SchemaProps: spec.SchemaProps{
+							Default:   "",
+							MinLength: ptr.To[int64](1),
+							MaxLength: ptr.To[int64](5),
+							Pattern:   "^a.*b$",
+							Type:      []string{"string"},
+							Format:    "",
+						},
+					},
+					"NumberValue": {
+						SchemaProps: spec.SchemaProps{
+							Default:          0,
+							Minimum:          ptr.To[float64](1),
+							Maximum:          ptr.To[float64](5),
+							ExclusiveMinimum: true,
+							ExclusiveMaximum: true,
+							MultipleOf:       ptr.To[float64](2),
+							Type:             []string{"number"},
+							Format:           "double",
+						},
+					},
+					"ArrayValue": {
+						SchemaProps: spec.SchemaProps{
+							MinItems:    ptr.To[int64](1),
+							MaxItems:    ptr.To[int64](5),
+							UniqueItems: true,
+							Type:        []string{"array"},
+							Items: &spec.SchemaOrArray{
+								Schema: &spec.Schema{
+									SchemaProps: spec.SchemaProps{
+										Default: "",
+										Type:    []string{"string"},
+										Format:  "",
+									},
+								},
+							},
+						},
+					},
+					"MapValue": {
+						SchemaProps: spec.SchemaProps{
+							MinProperties: ptr.To[int64](1),
+							MaxProperties: ptr.To[int64](5),
+							Type:          []string{"object"},
+							AdditionalProperties: &spec.SchemaOrBool{
+								Allows: true,
+								Schema: &spec.Schema{
+									SchemaProps: spec.SchemaProps{
+										Default: "",
+										Type:    []string{"string"},
+										Format:  "",
+									},
+								},
+							},
+						},
+					},
+					"celField": {
+						VendorExtensible: spec.VendorExtensible{
+							Extensions: spec.Extensions{
+								"x-kubernetes-validations": []interface{}{map[string]interface{}{"message": "string message", "rule": "self.length() > 0"}, map[string]interface{}{"messageExpression": "self + ' hello'", "rule": "self.length() % 2 == 0"}},
+							},
+						},
+						SchemaProps: spec.SchemaProps{
+							Default: "",
+							Type:    []string{"string"},
+							Format:  "",
+						},
+					},
+				},
+				Required: []string{"StringValue", "NumberValue", "ArrayValue", "MapValue"},
+			},
+			VendorExtensible: spec.VendorExtensible{
+				Extensions: spec.Extensions{
+					"x-kubernetes-validations": []interface{}{map[string]interface{}{"message": "foo", "rule": "self == oldSelf"}},
+				},
+			},
+		},
+	}
+}
+
+func schema_test_integration_testdata_valuevalidation_Foo2(ref common.ReferenceCallback) common.OpenAPIDefinition {
+	return common.EmbedOpenAPIDefinitionIntoV2Extension(valuevalidation.Foo2{}.OpenAPIV3Definition(), common.OpenAPIDefinition{
+		Schema: spec.Schema{
+			SchemaProps: spec.SchemaProps{
+				Description:   "This one has an open API v3 definition",
+				Type:          valuevalidation.Foo2{}.OpenAPISchemaType(),
+				Format:        valuevalidation.Foo2{}.OpenAPISchemaFormat(),
+				MaxProperties: ptr.To[int64](5),
+			},
+			VendorExtensible: spec.VendorExtensible{
+				Extensions: spec.Extensions{
+					"x-kubernetes-validations": []interface{}{map[string]interface{}{"message": "foo2", "rule": "self == oldSelf"}},
+				},
+			},
+		},
+	})
+}
+
+func schema_test_integration_testdata_valuevalidation_Foo3(ref common.ReferenceCallback) common.OpenAPIDefinition {
+	return common.EmbedOpenAPIDefinitionIntoV2Extension(common.OpenAPIDefinition{
+		Schema: spec.Schema{
+			SchemaProps: spec.SchemaProps{
+				Description:   "This one has a OneOf",
+				OneOf:         common.GenerateOpenAPIV3OneOfSchema(valuevalidation.Foo3{}.OpenAPIV3OneOfTypes()),
+				Format:        valuevalidation.Foo3{}.OpenAPISchemaFormat(),
+				MaxProperties: ptr.To[int64](5),
+			},
+			VendorExtensible: spec.VendorExtensible{
+				Extensions: spec.Extensions{
+					"x-kubernetes-validations": []interface{}{map[string]interface{}{"message": "foo3", "rule": "self == oldSelf"}},
+				},
+			},
+		},
+	}, common.OpenAPIDefinition{
+		Schema: spec.Schema{
+			SchemaProps: spec.SchemaProps{
+				Description:   "This one has a OneOf",
+				Type:          valuevalidation.Foo3{}.OpenAPISchemaType(),
+				Format:        valuevalidation.Foo3{}.OpenAPISchemaFormat(),
+				MaxProperties: ptr.To[int64](5),
+			},
+			VendorExtensible: spec.VendorExtensible{
+				Extensions: spec.Extensions{
+					"x-kubernetes-validations": []interface{}{map[string]interface{}{"message": "foo3", "rule": "self == oldSelf"}},
+				},
+			},
+		},
+	})
+}
+
+func schema_test_integration_testdata_valuevalidation_Foo5(ref common.ReferenceCallback) common.OpenAPIDefinition {
+	return common.EmbedOpenAPIDefinitionIntoV2Extension(valuevalidation.Foo5{}.OpenAPIV3Definition(), common.OpenAPIDefinition{
+		Schema: spec.Schema{
+			SchemaProps: spec.SchemaProps{
+				Type:          valuevalidation.Foo5{}.OpenAPISchemaType(),
+				Format:        valuevalidation.Foo5{}.OpenAPISchemaFormat(),
+				MinProperties: ptr.To[int64](1),
+				MaxProperties: ptr.To[int64](5),
+			},
+			VendorExtensible: spec.VendorExtensible{
+				Extensions: spec.Extensions{
+					"x-kubernetes-validations": []interface{}{map[string]interface{}{"message": "foo5", "rule": "self == oldSelf"}},
+				},
+			},
+		},
+	})
 }
